Fix remove button timeout
If the user clicked on remove all button and confirmed and then repeated
this action multiple times in quick succession, since the timeout wasn't
cleared it would quickly reset the state of the button even when the
button had already confirmed removal. Fixed by clearing previous timeout
handle if it exists.