Convert video list from object to array, implement search.
In order to support ordering, the displayed video list is now converted to an
array, whereas previously it was an object. There is hence now a
separation between the source of truth (which remains an object for fast
access of individual videos), and the displayed list (the array of
videos), which is dependent on the object. Search filtering is also
implemented, filtering from the videos array. Search bar was
implemented for testing, but is yet to be stylized.