Suppress uncaught promise errors for message send
New error introduced when Chrome updated API to utilize promises,
resulting in uncaught promise errors when sending attempting to send
a message to the popup when popup is not open. Previously this wasn't
an issue since a message to non-existing popup would simply be ignored.

Currently a workaround solution to suppress via downgrading the error to
a simple console log. A cleaner solution for the future is to figure out
how to detect when the popup is open and not send the message in the
first place when popup isn't open.