Fix: video undefined, overlooked error from 9f1b7b
When current url is YT home page, there is no video element, so checking
for the video element's data will cause an error. Fixed by requiring
handles to the video element to be created on-the-fly instead of reusing
a 'global' handle created on content script load, as well as checking if
video is defined before accessing its data.

The on-the-fly requirement is needed since content script is not
reloaded when navigating from YT home page -> video page. Thus, a global
handle won't work when navigating to video page as handle will remain
undefined. See d733d for this issue.